app/requirements.txt
autopep8==1.6.0
click==8.0.4
Deprecated==1.2.13
Flask==2.0.3
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
paho-mqtt==1.6.1
pycodestyle==2.8.0
pyparsing==3.0.7
redis==4.1.4
toml==0.10.2
Werkzeug==2.0.3
wrapt==1.13.3
zope.event==4.5.0
zope.interface==5.4.0
